requirements211119.txt
alembic==0.9.9
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
click==6.7
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.2
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
fpdf==1.7.2
gunicorn==19.9.0
httplib2==0.11.3
idna==2.6
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.15.1
Jinja2==2.10
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
lazy==1.3
Mako==1.0.7
MarkupSafe==1.0
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
oauth2==1.9.0.post1
oauthlib==2.0.7
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Pygments==2.4.2
pyrsistent==0.15.4
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.5
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.18.4
requests-oauthlib==0.8.0
Send2Trash==1.5.0
six==1.11.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.22
URLObject==2.4.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wincertstore==0.2
WTForms==2.1
WTForms-SQLAlchemy==0.1
zipp==0.6.0
